greeting: introducing (SIB) at startup
so you know what device you have, I guess.

I don't think there's a way to get the device signature or
name (eg. attiny414) because that requires reading off the target
memory. (SIB) is simple in that it's just part of the UDPI interface,
and should be usable ragardless of which target is connected.

and now since we need a proper serial connection, let's
run (updi-break) up front. if this break things for users, it's
possible to disable the greeting.