wip: updi link: send double break instead of single
ref https://github.com/kristianlm/updi-gdbserver/issues/1:

It appears that updi-gdbserver works once after running avrdude. So I
think avrdude initializes the target device correctly and leaves the
connection "ready" as it exits. updi-gdbserver leaves the target in a
"confused" state.

This patch hardcodes a double-break into the initialization
procedure. This is inefficient but seems to solve the problem for
me. Let's see if it also solves the issue 1's original poster's
problem too.