updi link: introduce updi-init
we add a timeout callback parameter, which defaults to the usual
error. however, a new updi-init procedure uses this to detect link
problems and issue a double-break as needed.

the double-break procedure seems to be able to bring the link back
online. the double-break is documentet in the UPDI sections of the
datasheets. I've seen it needed at irregular intervals, but ofter
after running avrdude. Probably since avrdude leaves the target with
the "UPDI disable" bit on:

avrdude: STCS 0x0C to address 0x03

This should hopefully make updi-gdbserver more robust.